+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Firmware: Support for Nintendo Wii Remote Motion Plus controller
+- Firmware: Support for Nintendo Wii Remote in vertical mode.
+
 ## [0.3] - 2019-08-05
 ### Added
 - Firmware: Nintendo Wii controller LED's represent the joystick port assigned to.