+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.4-rc0] - 2019-08-12
 ### Added
 - Firmware: Support for Nintendo Wii Remote Motion Plus controller.
-- Firmware: Support for "vertical mode" in Nintendo Wii Remote.
 - Firmware: Support for "accelerometer mode" in Nitendo Wii Remote.
+- Firmware: Support for "vertical mode" in Nintendo Wii Remote.
 - Docs: Improved "supported device" page.
 
 ## [0.3] - 2019-08-05